update packaging and perps v3 keeper
