+++ .dockerignore
@@ -20,3 +20,5 @@ coverage.xml
 .mypy_cache
 .pytest_cache
 .hypothesis
+.venv
+.silverback-sessions
