+++ docker-compose.yml
@@ -1,116 +1,34 @@
-version: "3"
-
 services:
-  order-keeper-arb:
+  keeper-base-mainnet:
     restart: always
     build: .
-    env_file: .env_arb
+    env_file: .env.base
     entrypoint:
       [
+        "uv",
+        "run",
         "silverback",
         "run",
-        "arbitrum_orders:app",
-        "--network",
-        "arbitrum:mainnet:alchemy",
-        "--runner",
-        "silverback.runner:WebsocketRunner",
-      ]
-
-  liq-keeper-arb:
-    restart: always
-    build: .
-    env_file: .env_arb
-    entrypoint:
-      [
-        "silverback",
-        "run",
-        "liquidations:app",
-        "--network",
-        "arbitrum:mainnet:alchemy",
-        "--runner",
-        "silverback.runner:WebsocketRunner",
-      ]
-
-  # order-keeper-arb-sepolia:
-  #   restart: always
-  #   build: .
-  #   env_file: .env_order_arb_sepolia
-  #   entrypoint: ["silverback", "run", "testnet:app", "--network", "arbitrum:sepolia:alchemy", "--runner", "silverback.runner:WebsocketRunner"]
-
-  # order-keeper-sepolia:
-  #   restart: always
-  #   build: .
-  #   env_file: .env_order_sepolia
-  #   entrypoint: ["silverback", "run", "testnet:app", "--network", "base:sepolia:alchemy", "--runner", "silverback.runner:WebsocketRunner"]
-
-  # liq-keeper-sepolia:
-  #   restart: always
-  #   build: .
-  #   env_file: .env_liq_sepolia
-  #   entrypoint: ["silverback", "run", "liquidations:app", "--network", "base:sepolia:alchemy", "--runner", "silverback.runner:WebsocketRunner"]
-
-  # order-keeper-sepolia:
-  #   restart: always
-  #   build: .
-  #   env_file: .env_order_sepolia
-  #   entrypoint:
-  #     [
-  #       "silverback",
-  #       "run",
-  #       "testnet:app",
-  #       "--network",
-  #       "base:sepolia:alchemy",
-  #       "--runner",
-  #       "silverback.runner:WebsocketRunner",
-  #     ]
-
-  # liq-keeper-sepolia:
-  #   restart: always
-  #   build: .
-  #   env_file: .env_liq_sepolia
-  #   entrypoint:
-  #     [
-  #       "silverback",
-  #       "run",
-  #       "liquidations:app",
-  #       "--network",
-  #       "base:sepolia:alchemy",
-  #       "--runner",
-  #       "silverback.runner:WebsocketRunner",
-  #     ]
-
-  liq-keeper-mainnet:
-    restart: always
-    build: .
-    env_file: .env
-    entrypoint:
-      [
-        "silverback",
-        "run",
-        "liquidations:app",
+        "v3_perps:app",
         "--network",
         "base:mainnet:alchemy",
         "--runner",
         "silverback.runner:WebsocketRunner",
       ]
 
-  order-keeper-mainnet:
+  keeper-arb-mainnet:
     restart: always
     build: .
-    env_file: .env
+    env_file: .env.arbitrum
     entrypoint:
       [
+        "uv",
+        "run",
         "silverback",
         "run",
-        "main:app",
+        "v3_perps:app",
         "--network",
-        "base:mainnet:alchemy",
+        "arbitrum:mainnet:alchemy",
         "--runner",
         "silverback.runner:WebsocketRunner",
       ]
-
-  # price-keeper-mainnet:
-  #   restart: always
-  #   build: .
-  #   env_file: .env.prices
-  #   entrypoint: ["silverback", "run", "price_pusher:app", "--network", "base:mainnet:alchemy", "--runner", "silverback.runner:WebsocketRunner"]
