+++ pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "synthetix-keepers"
+version = "0.1.0"
+description = "A collection of keepers for Synthetix contracts"
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = [
+    "eth-ape==0.8.15",
+    "python-dotenv>=1.0.1",
+    "silverback==0.5.8",
+    "synthetix>=0.1.20",
+]
